+++ config/app.php
@@ -237,5 +237,5 @@ return [
     | The only modification by laravel config
     |
     */
-    'version' => '1.0.2'
+    'version' => '1.0.3'
 ];